+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Backtrace Unity Release Notes
 
+## Version 3.0.2
+- `BacktraceDatabase` now provides a new `Send` method. This method will try to send all objects from the database respecting the client side deduplication and retry setting. This can be used as an alternative to the `Flush` method which will try to send all objects from the database ignoring any client side deduplication and retry settings.
+- `BacktraceClient` has been optimized to only serialize data as needed.
+- `BacktraceDatabase` `AutoSend` function has been optimized for performance improvements.
+- `BacktraceClient` by default will generate configuration file with client rate limit equal to 50.
+- Fixed invalid meta file.
+
 ## Version 3.0.1
 - The `BacktraceDatabase` class will now create database directory before final database validation. Previously, when directory didn't exist, BacktraceDatabase was disabled.
 - The `BacktraceDatabase` field now allows users to pass interpolated string in Database options. Developer can use `${Application.dataPath}` or `${Application.persistentDataPath}` to set path to database.